+++ main.py
@@ -1,10 +1,8 @@
+from fastapi import FastAPI
 from starlette.staticfiles import StaticFiles
-from starlette.templating import Jinja2Templates
-from fastapi import FastAPI, Request
-from pathlib import Path
 
-from routers import ChatbotHandlers, PlanController
-from routers import UserController
+from pathlib import Path
+from routers import (HomeController, ChatbotController, PlanController, UserController, HomeController)
 from dependencies.db import create_db
 
 app = FastAPI()
@@ -15,24 +13,12 @@ app.mount(
     name="static"
 )
 
-templates = Jinja2Templates(directory="templates")
-
-@app.get("/")
-async def read_root(request: Request):
-    return templates.TemplateResponse(
-        'home.html', {"request": request})
-
-@app.get("/chat")
-async def chat_page(request: Request):
-    return templates.TemplateResponse("chat.html", {"request": request})
-
-app.include_router(ChatbotHandlers.router)
-
 @app.on_event("startup")
 def on_startup():
     create_db()
 
-
+app.include_router(HomeController.router)
+app.include_router(ChatbotController.router)
 app.include_router(PlanController.router)
 app.include_router(UserController.router)
