+++ templates/home.html
@@ -31,6 +31,6 @@
         </section>
     </main>
 
-    <script src="/static/script.js"></script>
+    <script src="/static/home.js"></script>
 </body>
 </html>
